+++ common/src/androidTest/java/com/microsoft/identity/common/internal/platform/AndroidDevicePoPManagerEncryptionTests.java
@@ -22,10 +22,11 @@
 //  THE SOFTWARE.
 package com.microsoft.identity.common.internal.platform;
 
+import android.os.Build;
+
 import androidx.test.core.app.ApplicationProvider;
 
 import com.microsoft.identity.common.java.crypto.IDevicePopManager;
-import com.microsoft.identity.common.java.crypto.SigningAlgorithm;
 import com.microsoft.identity.common.java.exception.ClientException;
 
 import org.junit.After;
@@ -39,30 +40,49 @@ import java.io.IOException;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.CertificateException;
-import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.List;
 
+import static com.microsoft.identity.common.java.crypto.IDevicePopManager.Cipher.RSA_ECB_OAEPWithSHA_1AndMGF1Padding;
+import static com.microsoft.identity.common.java.crypto.IDevicePopManager.Cipher.RSA_ECB_OAEPWithSHA_256AndMGF1Padding;
+import static com.microsoft.identity.common.java.crypto.IDevicePopManager.Cipher.RSA_ECB_PKCS1_PADDING;
+import static com.microsoft.identity.common.java.crypto.IDevicePopManager.Cipher.RSA_NONE_OAEPWithSHA_1AndMGF1Padding;
 
 // Note: Test cannot use robolectric due to the following open issue
 // https://github.com/robolectric/robolectric/issues/1518
 //todo: Investigate if these tests can be migrated to common4j
 @RunWith(Parameterized.class)
-public class DevicePoPManagerSigningTests {
+public class AndroidDevicePoPManagerEncryptionTests {
 
-    private static final String DATA_TO_SIGN = "The quick brown fox jumped over the lazy dog.";
+    private static final String DATA_TO_ENCRYPT = "The quick brown fox jumped over the lazy dog.";
 
     private final IDevicePopManager devicePopManager;
-    private final SigningAlgorithm signingAlg;
+    private final IDevicePopManager.Cipher cipher;
 
     @Parameterized.Parameters
-    public static Iterable<SigningAlgorithm> testParams() {
-        return Arrays.asList(SigningAlgorithm.values());
+    public static Iterable<IDevicePopManager.Cipher> testParams() {
+        final List<IDevicePopManager.Cipher> ciphers = new ArrayList<>();
+
+        // Only execute these tests at appropriate API levels...
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
+            ciphers.add(RSA_ECB_PKCS1_PADDING);
+        }
+
+        //https://stackoverflow.com/questions/36015194/android-keystoreexception-unknown-error -
+        //https://issuetracker.google.com/issues/37075898
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            ciphers.add(RSA_NONE_OAEPWithSHA_1AndMGF1Padding);
+            ciphers.add(RSA_ECB_OAEPWithSHA_1AndMGF1Padding);
+            ciphers.add(RSA_ECB_OAEPWithSHA_256AndMGF1Padding);
+        }
+
+        return ciphers;
     }
 
-    @SuppressWarnings("unused")
-    public DevicePoPManagerSigningTests(final SigningAlgorithm signingAlg)
+    public AndroidDevicePoPManagerEncryptionTests(final IDevicePopManager.Cipher cipher)
             throws CertificateException, NoSuchAlgorithmException, KeyStoreException, IOException {
         devicePopManager = new AndroidDevicePopManager(ApplicationProvider.getApplicationContext());
-        this.signingAlg = signingAlg;
+        this.cipher = cipher;
     }
 
     @Before
@@ -76,12 +96,11 @@ public class DevicePoPManagerSigningTests {
     }
 
     @Test
-    public void testSigning() throws ClientException {
-        final String sampleSignature = devicePopManager.sign(signingAlg, DATA_TO_SIGN);
-        final boolean verified = devicePopManager.verify(signingAlg, DATA_TO_SIGN, sampleSignature);
+    public void testEncryption() throws ClientException {
+        final String cipherText = devicePopManager.encrypt(cipher, DATA_TO_ENCRYPT);
+        final String decryptedValue = devicePopManager.decrypt(cipher, cipherText);
 
-        if (!verified) {
-            Assert.fail("Failed signing/verification on: " + signingAlg);
-        }
+        Assert.assertEquals(DATA_TO_ENCRYPT, decryptedValue);
+        Assert.assertNotEquals(DATA_TO_ENCRYPT, cipherText);
     }
 }
